requirements.txt
fastapi
sqlalchemy
asyncpg
uvicorn

numpy
pandas
matplotlib
Pillow

pytest
pytest-asyncio
pytest-mock

pre-commit
flake8
isort
black
